Split AdministrationController
Moves the page related functions to dedicated controllers.
