+++ laravel/app/Http/Controllers/Administration/PhpInfoController.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Http\Controllers\Administration;
+
+use App\Http\Controllers\Controller;
+use Illuminate\View\View;
+
+class PhpInfoController extends Controller
+{
+    /**
+     * Display PHP information.
+     */
+    public function php_info(): View
+    {
+        return view('administration.phpinfo');
+    }
+}
